requirements.txt
attrs==19.3.0
autopep8==1.5.2
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
decorator==4.4.2
Flask==1.1.1
fuzzywuzzy==0.18.0
GeoAlchemy2==0.6.3
google==2.0.3
google-search==1.0.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
pathlib-mate==1.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3
requests==2.23.0
requests-file==1.5.1
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
tldextract==2.2.2
urllib3==1.25.9
uszipcode==0.2.4
validators==0.15.0
Werkzeug==1.0.0
